% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_elements}
\alias{add_elements}
\title{Add elements}
\usage{
add_elements(x, desired_length, element = NA)
}
\arguments{
\item{x}{Vector.}

\item{desired_length}{Integer.}

\item{element}{The element to add at the end of \code{x}.}
}
\value{
A vector of length \code{desired_length}.
}
\description{
Add repeted elements up to a desired length.
}
\examples{
scrapeacexchange:::add_elements(1:4, 7)
}
