% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_dep.R
\name{check_nodejs}
\alias{check_nodejs}
\title{Check if Node.js is installed}
\usage{
check_nodejs()
}
\value{
Logical
}
\description{
Check if Node.js is installed
}
\examples{
harvestacexchange:::check_nodejs()
}
