% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_turnips}
\alias{clean_turnips}
\title{Clean harvested turnips prices}
\usage{
clean_turnips(text)
}
\arguments{
\item{text}{Character of length 5 or 6}
}
\value{
Character of length 6.
}
\description{
Clean harvested turnips prices
}
