% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_old_turnips}
\alias{filter_old_turnips}
\alias{filter_old_events}
\title{Filter a dataframe based on another one}
\usage{
filter_old_turnips(df, df_old = NULL)

filter_old_events(df, df_old = NULL)
}
\arguments{
\item{df}{Formatted data frame.}

\item{df_old}{A dataframe of the same type to be compared to.}
}
\value{
A filtered dataframe.
}
\description{
Filter a dataframe based on another one
}
