% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_price_min}
\alias{filter_price_min}
\alias{filter_grade_min}
\alias{filter_line}
\alias{filter_no_tag}
\alias{filter_words}
\title{Filter functions}
\usage{
filter_price_min(df, price_min = 500)

filter_grade_min(df, grade_min = 75, keep_new = TRUE)

filter_line(df)

filter_no_tag(df)

filter_words(df, words)
}
\arguments{
\item{df}{Formatted data frame.}

\item{price_min}{Minimum price.}

\item{grade_min}{Minimum grade.}

\item{keep_new}{Logical. Should host without grade should be kept?}

\item{words}{Words to find in notice column.}
}
\value{
A filtered dataframe.
}
\description{
Filter functions
}
\examples{
\dontrun{
turnips \%>\%
  filter_price_min(price_min = 400) \%>\%
  filter_grade_min(grade_min = 90, keep_new = TRUE) \%>\%
  filter_line() \%>\%
df_events \%>\%
  filter_words(words = c("gold", "celeste")) \%>\%
  filter_no_tag()
}

}
