% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_tibble.R
\name{html_to_tibble_turnips}
\alias{html_to_tibble_turnips}
\alias{html_to_tibble_events}
\title{Convert HTML files to tibbles}
\usage{
html_to_tibble_turnips(path_to_html)

html_to_tibble_events(path_to_html)
}
\arguments{
\item{path_to_html}{An html file.}
}
\value{
A tibble
}
\description{
Convert HTML files to tibbles
}
