% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_and_write}
\alias{scrape_and_write}
\title{Scrape webpage}
\usage{
scrape_and_write(page, path)
}
\arguments{
\item{page}{The targeted URL.}

\item{path}{Path to write to.}
}
\value{
\code{NULL}
}
\description{
Scrape webpage
}
