% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{turnips}
\alias{turnips}
\alias{events}
\title{Harvest turnips prices and events}
\usage{
turnips()

events()
}
\value{
A tibble.
}
\description{
Harvest turnips prices and events
}
\examples{
\dontrun{
turnips()
events()
}
}
