% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\docType{data}
\name{url_turnips}
\alias{url_turnips}
\alias{url_events}
\title{URL to animalcrossingexchange pages}
\format{
Character

An object of class \code{character} of length 1.
}
\usage{
url_turnips

url_events
}
\description{
URL to animalcrossingexchange pages
}
\examples{
url_turnips
url_events
}
\keyword{datasets}
