% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_events}
\alias{clean_events}
\title{Clean harvested events}
\usage{
clean_events(text)
}
\arguments{
\item{text}{Characterof length 9 or 11.}
}
\value{
Character of length 7.
}
\description{
Clean harvested events
}
