% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvest_and_write}
\alias{harvest_and_write}
\title{Harvest webpage}
\usage{
harvest_and_write(page, path_write)
}
\arguments{
\item{page}{The targeted URL.}

\item{path_write}{Path to write to.}
}
\value{
\code{NULL}
}
\description{
Harvest webpage
}
