% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_dep.R
\name{install_js_dep}
\alias{install_js_dep}
\title{Install JavaScript libraries and scripts}
\usage{
install_js_dep(dir_js = ".js")
}
\arguments{
\item{dir_js}{The directory where all JS dependencies will be installed.
Default to \code{.js}.}
}
\value{
\code{NULL}
}
\description{
Install JavaScript libraries and scripts
}
\details{
This must be done once per project. You must have Node.js installed
on your computer.
}
\examples{
\dontrun{install_js_dep()}
}
