% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{turnips}
\alias{turnips}
\alias{events}
\title{Harvest turnips prices and events}
\usage{
turnips(dir_js = ".js")

events(dir_js = ".js")
}
\arguments{
\item{dir_js}{The directory where all the JS dependencies are installed.
Default to \code{.js}.}
}
\value{
A tibble.
}
\description{
Harvest turnips prices and events
}
\examples{
\dontrun{
turnips()
events()
}
}
